Initial commit of burgeoning Cube class
This commit adds the naive implementation of the cube state using a
vector of vector of chars as well as a method to print the state and
place holders for most of the face rotation methods. The 'u' methods has
been implemented so far and the rest need to be written and tested.
There is a small main function for testing.