Add PermuteFace method
This commit adds a new method that generalizes the face permutation
logic. These permutations permute the cubies on the face of the face
being rotated, not the adjacent edges. I.e., if the right face is being
rotated, this method permutes the r colors/cubies, but none of the
d/f/u/b colors/cubies.