Generalize overloaded ==/equality operator
This commit removes the blaring assumption that both cubes being
compared with this operator were oriented the same way. Now, a temporary
copy of the right hand side is created and manipulated so that it
matches the orientation of the left hand side (lhs) cube. Then, this
temporary, re-oriented cube is compared against the lhs/self.