Implement rotation of the top and middle slices
This commit adds the logic required to rotate the top/up and horizontal
middle slices. Three wrapper functions are introduced here, the U, Up,
and UU methods. This is currently implemented by reusing both the
methods used to rotate the top/up face and the adjacent horizontal
middle slice in lieu of duplicating a lot of code.