Implement SetCubeState method
This commit introduces a method used to set a cube state; currently this
method takes as input a std::vector of chars that must be 54 elements
long. There is _no_ other validation performed. This needs a lot of
work, but is a decent first step to being able to input a cube state.