Add a couple tests for SetCubeState
This commit adds two tests (which I do not love) for the SetCubeState
method. I wanted to test that it throws an exception when supplied with
the incorrect number of cube elements but the EXPECT_THROW googletest
macro is confusing. I am not certain that these tests are really
complete nor that they even work the way I intend them to.