Improve IsSolvedWorksNegative googletest
This commit fixes an issue where this test succeeded when it was meant
to fail. Only one random scramble was being performed, and if it was a
cube rotation, the cube was still solved and this test breaks. For now,
I have simply increased the number of random moves used to scramble the
cube to 15 which will make this _much_ less likely to happen.